export_{pfc,bpf} take a &mut T where T: AsRawFd
Taking ownership of the output file is bad if you want to do something
with the output (e.g. display to the user or pass to an other program).
Usually you would use `export_bpf(file.try_clone()?)?` to workaround
this. (OT: And remeber to `file.rewind()?`)
However I think it is save to take a &mut.

Using T: AsRawFd instead of file allows to pass types like ChildStdout
as well.

In the future (when io_safety is stabelized), we can change this to take
BorrowedFd.