build: Set PKG_CONFIG_ALLOW_CROSS for the musl target
To allow pkg-config support cross-compilation for the musl target,
the build script sets `PKG_CONFIG_ALLOW_CROSS` environment variable to 1.

If the variable is not set, pkg-config cannot extract `PKG_CONFIG_PATH`
for the path of the libseccomp library compiled against the musl-libc
when users try to compile the libseccomp crate under the gnu-libc host,
so the crate cannot enable the `cfg` for `libseccomp_v2_5`.
As a result, the `notify.rs` will not be compiled.

Signed-off-by: Manabu Sugimoto <Manabu.Sugimoto@sony.com>