Fix misleading example
seccomp_syscall_resolve_name_arch resolves the given syscall name to the
syscall number (or pseudo number) it has on the give architectur.
If you are not running on this archtecture, the result isn't the native
number of the syscall.

seccomp_rule_add interprets the given syscall number for the native
archtecture. Passing a number resolved for a foreign architecture, gives
you any result but not what you want. In the best case this number is
invalid for you native architecture and a error is returend, but it is
also possible that the number is valid, then the blocked syscall is any
arbitrary syscall.

Signed-off-by: rusty-snake <41237666+rusty-snake@users.noreply.github.com>