Fix regression from fe7bd6a/#143
    let comperators: &[ScmpArgCompare] = &[scmp_cmp!($arg0 & 123 == 123)];

Used to work before fe7bd6a/#143 but fail now with "creates a temporary
which is freed while still in use". Refactoing the macro back to
CallExpression instead of BlockExpression as utmost token fixes it.
However IIRC there was a reason why I did not made it that way in the
first place and IIRC it wasn't working because attributes in that
position were unstable. This would effect MSRV (#135). Nevermind, I can
not find anything about this in rust/RELEASES.md or with an internet
search, neither my tries with older compilers made it fail that why.

Signed-off-by: rusty-snake <41237666+rusty-snake@users.noreply.github.com>