Fix check_version logic
If the major version of the linked libseccomp library is equal to or greater
than the specified version in the `check_version`, the function is successful.

The previous implementation makes the `check_version` fail if the major version
is not equal to the specified version, but this is not practical because
when the libseccomp library updates the major version, all the apis checked
by the function fail even though they are still compatible. As a result,
we'll have to change the all the apis at that time.

Signed-off-by: Manabu Sugimoto <Manabu.Sugimoto@sony.com>