doc: Show ScmpSyscall::new of const-syscall  with annotation
Allow the crates.io documentation to show `ScmpSyscall::new()`
provided by the `const-syscall` feature with the following annotation.

> Available on crate feature `const-syscall` only.

This is enabled by the `doc_cfg` feature which is
available in only the nightly channel, but the `docs.rs` of crates.io
is built by the nightly so we don't worry about it.
Ref. https://docs.rs/about/builds

Signed-off-by: Manabu Sugimoto <Manabu.Sugimoto@sony.com>